pyproject.toml
[project]
name = "pynebula"
version = "0.1.0"
description = "A Python wrapper for the Nebula API"
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "fastapi>=0.115.8",
    "langchain (>=0.3.17,<0.4.0)",
    "langchain-openai (>=0.3.3,<0.4.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "openai (>=1.61.0,<2.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
]
